🏷️🐛 [RUMF-1267] remove dependency loop in recorder types
This commit reorganize recorder types moving most of the
src/domain/record/types in src/types. I wanted to do this for a long
time, because I'm always confused when looking for a specific type of
the recorder. Now everything is at the same level.
