+++ packages/rum/src/entries/internal.ts
@@ -5,22 +5,6 @@
  * WARNING: this module is not intended for public usages, and won't follow semver for breaking
  * changes.
  */
-export {
-  MutationData,
-  MousemoveData,
-  MouseInteractionData,
-  ScrollData,
-  ViewportResizeData,
-  InputData,
-  MediaInteractionData,
-  StyleSheetRuleData,
-  MediaInteractions,
-  MouseInteractions,
-  AddedNodeMutation,
-  MousePosition,
-  RemovedNodeMutation,
-} from '../domain/record'
-
 export { PRIVACY_ATTR_NAME, PRIVACY_ATTR_VALUE_HIDDEN, PRIVACY_CLASS_HIDDEN, NodePrivacyLevel } from '../constants'
 
 export * from '../types'
