+++ packages/rum/src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './segment'
+export * from './record'
+export * from './serializedNode'
