+++ test/e2e/scenario/recorder/recorder.scenario.ts
@@ -1,7 +1,5 @@
-import type { CreationReason, Segment } from '@datadog/browser-rum/src/types'
-import { IncrementalSource, RecordType } from '@datadog/browser-rum/src/types'
-import type { InputData, StyleSheetRuleData } from '@datadog/browser-rum/src/domain/record/types'
-import { NodeType } from '@datadog/browser-rum/src/domain/record/types'
+import type { InputData, StyleSheetRuleData, CreationReason, Segment } from '@datadog/browser-rum/src/types'
+import { NodeType, IncrementalSource, RecordType } from '@datadog/browser-rum/src/types'
 
 import type { RumInitConfiguration } from '@datadog/browser-rum-core'
 import { DefaultPrivacyLevel } from '@datadog/browser-rum'
